requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
transformers==4.33.1
torch==2.0.1
pydantic==2.3.0
python-dotenv==1.0.0
